feat(Action) Adding updateCurrentStamina action
This commit adds in a updateCurrentStamina redux action, and also
updates the staminaGraph to display a more dynamic line graph.

Right now the line graph simply draws a line from the current stamina,
to whatever point the stamina reaches its maximum, and it flatlines
until the end. I'm going to make it a little more dynamic when I start
adding in "events"